src/tests/forests_test.py
import numpy as np
import jax.numpy as jnp
from jax import jit, grad
import networkx as nx
from forests import kruskals, pairwise_square_distance, mwst, kruskals_constrained, mwst_constrained
from forests import pfkb
from scipy.spatial import distance_matrix
np.random.seed(0)


def sample_and_check_pwdist(npoints, d):
    X = np.random.randn(npoints, d)
    # scipy pw distance matrix
    SD = distance_matrix(X, X) ** 2
    # jax pw distance matrix
    JD = pairwise_square_distance(X)
    return np.all(np.isclose(SD, JD))

def test_trivial_pwdist():
    X = np.eye(100)
    SD = distance_matrix(X, X) ** 2
    JD = pairwise_square_distance(X)
    assert np.all(np.isclose(SD, JD))

def test_pairwise_square_distance():
    """
    There is a numerical difference between our pw distance matrix
    and that that is in scipy. Possible because jax is using a lower
    floating point preicsion? But if d=1 there can be some differences
    in the calculated distances
    """
    # assert sample_and_check_pwdist(100, 1) # this one causes problems
    assert sample_and_check_pwdist(10, 2)
    assert sample_and_check_pwdist(50, 4)
    assert sample_and_check_pwdist(10, 40)


def make_edge_dict(A, D):
    rowids, colids = np.where(A  - np.eye(A.shape[0]) > 0 )
    E = {key : D[key[0], key[1]] for key in zip(rowids, colids)}
    return E


def sample_and_check_kruskal(npoints, d):
    # sample data and calculate pw dist matrix
    X = np.random.randn(npoints, d)
    n = len(X)
    D = pairwise_square_distance(X)
    S = - D

    # calculate A using kruskals
    A, M = kruskals(D)

    # calculate hopefully the same matricies using maximum weight spanning tree of S
    A_, M_ = mwst(S)

    # calculate A again using networkx
    G = nx.Graph()
    EG = make_edge_dict(np.ones((n, n)), D)
    for key, value in EG.items():
        G.add_edge(key[0], key[1], weight=value)
    TG = nx.minimum_spanning_tree(G)
    TA = nx.adjacency_matrix(TG).toarray()

    # check both results are the same
    assert (A == A_).all()
    assert (M == M_).all()

    mst = (A - jnp.eye(n)) * D
    assert np.isclose(TA, mst).all()


def test_kruskals():
    sample_and_check_kruskal(10, 2)
    sample_and_check_kruskal(50, 3)
    sample_and_check_kruskal(100, 1)
    sample_and_check_kruskal(20, 1000)
    sample_and_check_kruskal(2, 1)


def test_kruskals_simple():

    D = jnp.array([
        [0., 4, 5],
        [4., 0, 3],
        [5., 3, 0]
    ])

    # A0, M0 = kruskals(D, ncc=0)
    # A1, M1 = kruskals(D, ncc=1)
    # A2, M2 = kruskals(D, ncc=2)

    # 3 connected components --> forest with no edges
    A3, M3 = kruskals(D, ncc=3)
    A3_, M3_ = mwst(-D, ncc=3)
    assert np.isclose(A3, A3_).all()
    assert np.isclose(M3, M3_).all()
    assert np.isclose(A3, np.eye(3)).all()
    assert np.isclose(M3, np.eye(3)).all()

    # 2 connected components links node 2 and 3
    A2, M2 = kruskals(D, ncc=2)
    A2_, M2_ = mwst(-D, ncc=2)
    A2star = np.array([
        [1, 0, 0],
        [0, 1, 1],
        [0, 1, 1]
    ])
    assert np.isclose(A2, A2_).all()
    assert np.isclose(M2, M2_).all()
    assert np.isclose(A2, A2star).all()
    # M2star = A2star
    assert np.isclose(M2, A2star).all()

    #one  connected component implies full link
    A1, M1 = kruskals(D, ncc=1)
    A1_, M1_ = mwst(-D, ncc=1)
    A1star = np.array([
        [1, 1, 0],
        [1, 1, 1],
        [0, 1, 1]
    ])
    assert np.isclose(A1, A1_).all()
    assert np.isclose(M1, M1_).all()
    assert np.isclose(A1, A1star).all()
    assert np.isclose(M1, np.ones((3, 3))).all()


@jit
def standardize(D):
    return (D - D.mean()) / D.std()

def random_standarize_invariance_check(n, d):
    X = np.random.randn(n, d)
    D = pairwise_square_distance(X)
    D_std = standardize(D)

    A1, M1 = kruskals(D, ncc=1)
    A2, M2 = kruskals(D, ncc=2)

    A1_std, M1_std = kruskals(D_std, ncc=1)
    A2_std, M2_std = kruskals(D_std, ncc=2)

    assert jnp.isclose(A1, A1_std).all()
    assert jnp.isclose(A2, A2_std).all()
    assert jnp.isclose(M1, M1_std).all()
    assert jnp.isclose(M2, M2_std).all()



def test_kruskals_scale_invariance():
    random_standarize_invariance_check(10, 3)
    random_standarize_invariance_check(5, 1000)
    random_standarize_invariance_check(50, 150)


def test_constrained_example():

    X = np.array([
           [ 0.10957423, -0.09405555],
           [ 0.40508325, -0.66398068],
           [-0.34922997, -0.82562293],
           [-0.29366213,  0.87128886],
           [-0.44979371, -0.18191248]])

    D = pairwise_square_distance(X)
    C_star = np.zeros_like(X @ X.T)
    C_star[2, 0] = -1
    C_star[0, 2] = -1
    A, M = kruskals_constrained(D, C_star, ncc=2)

    M_true = np.array(
       [[1., 1., 0., 1., 1.],
       [1., 1., 0., 1., 1.],
       [0., 0., 1., 0., 0.],
       [1., 1., 0., 1., 1.],
       [1., 1., 0., 1., 1.]]
    )


    A_true = np.array([
       [1., 1., 0., 1., 1.],
       [1., 1., 0., 0., 0.],
       [0., 0., 1., 0., 0.],
       [1., 0., 0., 1., 0.],
       [1., 0., 0., 0., 1.]]
    )

    assert jnp.isclose(M_true, M).all()
    assert jnp.isclose(A_true, A).all()





def test_simple_constraint():
    D = jnp.array([
        [0., 3., 1.],
        [3., 0., 2.],
        [1., 2., 0.],
    ])


    # M where the 3rd node is isolated
    C_star = jnp.array([
        [1., 1., -1],
        [1., 1., -1],
        [-1., -1., 1.],
    ])

    M_true = jnp.array([
        [1., 1., 0.],
        [1., 1., 0.],
        [0., 0., 1.],
    ])

    A_true = jnp.array([
        [1., 1., 0.],
        [1., 1., 0.],
        [0., 0., 1.]
    ])


    A, M = kruskals_constrained(D, C_star, ncc=2)

    assert jnp.isclose(M, M).all()
    assert jnp.isclose(A, A_true).all()

    A_, M_ = mwst_constrained(-D, C_star, ncc=2)
    assert jnp.isclose(A_, A_true).all()
    assert jnp.isclose(M_, M_true).all()



def test_blocking_join():

    N = 25
    Xr = np.random.randn(N, 2)
    Dr = pairwise_square_distance(Xr)

    M_target = np.eye(N)
    M_target[0, 1]+= 1
    M_target[1, 0]+= 1

    C_star = - np.ones_like(M_target) + 2 * np.eye(N)
    C_star[0, 1] = 1
    C_star[1, 0] = 1

    A, M = kruskals_constrained(Dr, C_star, ncc=N - 1)
    A_, M_ = mwst_constrained(-Dr, C_star, ncc=N - 1)

    assert jnp.isclose(A_, A).all()
    assert jnp.isclose(M, M_).all()
    assert jnp.isclose(M_target, M).all()
    assert jnp.isclose(M_target, A).all()



def test_blocking_join_bias_vs_unbias():

    N = 25
    Xr = np.random.randn(N, 2)
    Dr = pairwise_square_distance(Xr)

    M_target = np.eye(N)
    M_target[0, 1]+= 1
    M_target[1, 0]+= 1

    C_star = - np.ones_like(M_target) + 2 * np.eye(N)
    C_star[0, 1] = 1
    C_star[1, 0] = 1

    Ab, Mb = kruskals_constrained(Dr, C_star, ncc=N - 1, use_bias=True)
    Au, Mu = kruskals_constrained(Dr, C_star, ncc=N - 1, use_bias=False)

    assert jnp.isclose(Ab, Au).all()
    assert jnp.isclose(Mb, Mu).all()


def test_bias_unconstrained():
    """
    using bias or no bias should not affect anything for
    unconstrained kruskals
    """

    N = 25
    Xr = np.random.randn(N, 2)
    Dr = pairwise_square_distance(Xr)

    Ab, Mb = kruskals_constrained(Dr, np.zeros_like(Dr), ncc=10, use_bias=True)
    Au, Mu = kruskals_constrained(Dr, np.zeros_like(Dr), ncc=10, use_bias=False)

    assert jnp.isclose(Ab, Au).all()
    assert jnp.isclose(Mb, Mu).all()



